Alter how scopes are stored.
"Local" is now stored as "Global.Local" forcing the period as a
separator.  This has a few side effects:
(1) reuse of a local is possible since scopes are not absolutely unique
anymore, they are unique as far as their absolute address.
(2) A local scope should only define its local name.  Previously some
uses of this library were using A.B as the local name, now it should
provide B and this class will derive the "A." portion